ensureAuthorization 모듈화와 구현

+++ controller/LikeController.js
@@ -1,41 +1,63 @@
+const ensureAuthorization = require("../auth"); // 인증 모듈
+const jwt = require("jsonwebtoken");
 const conn = require("../mariadb");
 const { StatusCodes } = require("http-status-codes");
 
 const addLike = (req, res) => {
-
-    const { id } = req.params; //book_id
-    const { user_id } = req.body;
-
     // 좋아요 추가
-    let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
-    values = [user_id, id];
-    conn.query(sql, values, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+    const book_id = req.params.id;
 
-        return res.status(StatusCodes.OK).json(results);
-    });
+    let authorization = ensureAuthorization(req, res);
+    
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인해주세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        });
+    } else {
+        let sql = "INSERT INTO likes (user_id, liked_book_id) VALUES (?, ?);";
+        values = [authorization.id, book_id];
+        conn.query(sql, values, (err, results) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
+
+            return res.status(StatusCodes.OK).json(results);
+        });
+    }
 };
 
 const removeLike = (req, res) => {
     //좋아요 제거
-    const { id } = req.params; //book_id
-    const { user_id } = req.body;
+    const book_id = req.params.id;
 
-    let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
-    values = [user_id, id];
-    conn.query(sql, values, (err, results) => {
-        if (err) {
-            console.log(err);
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+    let authorization = ensureAuthorization(req, res);
+    
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message": "로그인 세션이 만료되었습니다. 다시 로그인해주세요."
+        });
+    } else if (authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message": "잘못된 토큰입니다."
+        });
+    } else {
+        let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?";
+        values = [authorization.id, book_id];
+        conn.query(sql, values, (err, results) => {
+            if (err) {
+                console.log(err);
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
 
-        return res.status(StatusCodes.OK).json(results);
-    });
+            return res.status(StatusCodes.OK).json(results);
+        });
+    }
 };
-
 
 module.exports = {
     addLike,
